#1016 review fixes for api calls and consistency
