--- frontend/src/enums/concept-label.enum.ts
@@ -1,6 +1,0 @@
-export enum ConceptLabelEnum {
-  PREF = 'prefLabel',
-  ALT = 'altLabel',
-  HIDDEN = 'hiddenLabel',
-  SORT = 'sortLabel',
-}
