--- frontend/src/enums/concept-note.enum.ts
@@ -1,9 +1,0 @@
-export enum ConceptNoteEnum {
-  CHANGE = 'changeNote',
-  DEFINITION = 'definition',
-  EDITORIAL = 'editorialNote',
-  EXAMPLE = 'example',
-  HISTORY = 'historyNote',
-  SCOPE = 'scopeNote',
-  NOTE = 'note',
-}
