+++ frontend/vite.config.ts
@@ -59,7 +59,6 @@ export default defineConfig({
       '@models': fileURLToPath(new URL('./src/models', import.meta.url)),
       '@enums': fileURLToPath(new URL('./src/enums', import.meta.url)),
       '@views': fileURLToPath(new URL('./src/views', import.meta.url)),
-      '@stores': fileURLToPath(new URL('./src/stores', import.meta.url)),
       '@': fileURLToPath(new URL('./src', import.meta.url)),
     },
   },
